@@ -5,6 +5,7 @@ import (
 	"net/http"
 
 	"github.com/google/uuid"
+	"github.com/keertirajmalik/expenser/expenser-server/auth"
 	"github.com/keertirajmalik/expenser/expenser-server/internal/model"
 	"github.com/keertirajmalik/expenser/expenser-server/logger"
 	"github.com/shopspring/decimal"
@@ -12,10 +13,15 @@ import (
 
 func HandleTransactionGet(transactionService model.TransactionService) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		userID := r.Context().Value("userID").(uuid.UUID)
+		userID, ok := auth.UserIDFromContext(r.Context())
+		if !ok {
+			respondWithError(w, http.StatusUnauthorized, "Unauthorized")
+			return
+		}
+
 		transactions, err := transactionService.GetTransactionsFromDB(r.Context(), userID)
 		if err != nil {
-			logger.Error("Error while fetching tranasactions", map[string]interface{}{
+			logger.Error("Error while fetching transactions", map[string]any{
 				"error": err,
 			})
 			respondWithError(w, http.StatusInternalServerError, err.Error())
@@ -40,7 +46,7 @@ func HandleTransactionCreate(transactionService model.TransactionService) http.H
 		params := parameters{}
 		err := decoder.Decode(&params)
 		if err != nil {
-			logger.Error("Error while decoding parameters", map[string]interface{}{
+			logger.Error("Error while decoding parameters", map[string]any{
 				"params": params,
 				"error":  err,
 			})
@@ -48,7 +54,11 @@ func HandleTransactionCreate(transactionService model.TransactionService) http.H
 			return
 		}
 
-		userID := r.Context().Value("userID").(uuid.UUID)
+		userID, ok := auth.UserIDFromContext(r.Context())
+		if !ok {
+			respondWithError(w, http.StatusUnauthorized, "Unauthorized")
+			return
+		}
 
 		transaction := model.InputTransaction{
 			ID:       uuid.New(),
@@ -65,7 +75,7 @@ func HandleTransactionCreate(transactionService model.TransactionService) http.H
 			respondWithError(w, http.StatusBadRequest, err.Error())
 			return
 		}
-		respondWithJson(w, http.StatusOK, dbTransaction)
+		respondWithJson(w, http.StatusCreated, dbTransaction)
 	}
 }
 
@@ -83,7 +93,7 @@ func HandleTransactionUpdate(transactionService model.TransactionService) http.H
 
 		id, err := uuid.Parse(idStr)
 		if err != nil {
-			logger.Error("Error while parsing transaction ID", map[string]interface{}{
+			logger.Error("Error while parsing transaction ID", map[string]any{
 				"error": err,
 				"uuid":  idStr,
 			})
@@ -95,7 +105,7 @@ func HandleTransactionUpdate(transactionService model.TransactionService) http.H
 		params := parameters{}
 		err = decoder.Decode(&params)
 		if err != nil {
-			logger.Error("Error while decoding parameters", map[string]interface{}{
+			logger.Error("Error while decoding parameters", map[string]any{
 				"error":  err,
 				"params": params,
 			})
@@ -103,7 +113,11 @@ func HandleTransactionUpdate(transactionService model.TransactionService) http.H
 			return
 		}
 
-		userID := r.Context().Value("userID").(uuid.UUID)
+		userID, ok := auth.UserIDFromContext(r.Context())
+		if !ok {
+			respondWithError(w, http.StatusUnauthorized, "Unauthorized")
+			return
+		}
 
 		transaction := model.InputTransaction{
 			ID:       id,
@@ -130,7 +144,7 @@ func HandleTransactionDelete(transactionService model.TransactionService) http.H
 
 		id, err := uuid.Parse(idStr)
 		if err != nil {
-			logger.Error("Error while parsing uuid", map[string]interface{}{
+			logger.Error("Error while parsing uuid", map[string]any{
 				"error": err,
 				"uuid":  idStr,
 			})
@@ -138,10 +152,14 @@ func HandleTransactionDelete(transactionService model.TransactionService) http.H
 			return
 		}
 
-		userID := r.Context().Value("userID").(uuid.UUID)
+		userID, ok := auth.UserIDFromContext(r.Context())
+		if !ok {
+			respondWithError(w, http.StatusUnauthorized, "Unauthorized")
+			return
+		}
 		err = transactionService.DeleteTransactionFromDB(r.Context(), id, userID)
 		if err != nil {
-			logger.Error("Error while deleting transaction", map[string]interface{}{
+			logger.Error("Error while deleting transaction", map[string]any{
 				"transaction_id": id,
 				"user_id":        userID,
 				"error":          err,
