@@ -5,13 +5,18 @@ import (
 	"net/http"
 
 	"github.com/google/uuid"
+	"github.com/keertirajmalik/expenser/expenser-server/auth"
 	"github.com/keertirajmalik/expenser/expenser-server/internal/model"
 	"github.com/keertirajmalik/expenser/expenser-server/logger"
 )
 
 func HandleCategoryGet(categoryService model.CategoryService) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		userID := r.Context().Value("userID").(uuid.UUID)
+		userID, ok := auth.UserIDFromContext(r.Context())
+		if !ok {
+			respondWithError(w, http.StatusUnauthorized, "Unauthorized")
+			return
+		}
 		categories, err := categoryService.GetCategoriesFromDB(r.Context(), userID)
 		if err != nil {
 			respondWithError(w, http.StatusInternalServerError, "Failed to retrieve categories")
@@ -41,7 +46,11 @@ func HandleCategoryCreate(categoryService model.CategoryService) http.HandlerFun
 			return
 		}
 
-		userID := r.Context().Value("userID").(uuid.UUID)
+		userID, ok := auth.UserIDFromContext(r.Context())
+		if !ok {
+			respondWithError(w, http.StatusUnauthorized, "Unauthorized")
+			return
+		}
 		category, err := categoryService.AddCategoryToDB(r.Context(), model.Category{
 			ID:          uuid.New(),
 			Name:        params.Name,
@@ -55,7 +64,7 @@ func HandleCategoryCreate(categoryService model.CategoryService) http.HandlerFun
 			return
 		}
 
-		respondWithJson(w, http.StatusOK, category)
+		respondWithJson(w, http.StatusCreated, category)
 	}
 }
 
@@ -92,7 +101,11 @@ func HandleCategoryUpdate(categoryService model.CategoryService) http.HandlerFun
 			return
 		}
 
-		userID := r.Context().Value("userID").(uuid.UUID)
+		userID, ok := auth.UserIDFromContext(r.Context())
+		if !ok {
+			respondWithError(w, http.StatusUnauthorized, "Unauthorized")
+			return
+		}
 		category, err := categoryService.UpdateCategoryInDB(r.Context(), model.Category{
 			ID:          id,
 			Name:        params.Name,
@@ -119,7 +132,11 @@ func HandleCategoryDelete(categoryService model.CategoryService) http.HandlerFun
 			return
 		}
 
-		userID := r.Context().Value("userID").(uuid.UUID)
+		userID, ok := auth.UserIDFromContext(r.Context())
+		if !ok {
+			respondWithError(w, http.StatusUnauthorized, "Unauthorized")
+			return
+		}
 		err = categoryService.DeleteCategoryFromDB(r.Context(), id, userID)
 		if err != nil {
 			respondWithError(w, http.StatusBadRequest, err.Error())
