@@ -1,31 +1,34 @@
 import AccountPage from "@/app/sidebar/account";
+import ImportTransactionPage from "@/app/sidebar/import-transaction";
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuGroup,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
+    DropdownMenu,
+    DropdownMenuContent,
+    DropdownMenuGroup,
+    DropdownMenuItem,
+    DropdownMenuLabel,
+    DropdownMenuSeparator,
+    DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 import {
-  SidebarMenu,
-  SidebarMenuButton,
-  SidebarMenuItem,
-  useSidebar,
+    SidebarMenu,
+    SidebarMenuButton,
+    SidebarMenuItem,
+    useSidebar,
 } from "@/components/ui/sidebar";
 import { Skeleton } from "@/components/ui/skeleton";
 import { useGetUserQuery } from "@/hooks/use-user-query";
 import { generateAvatarName } from "@/lib/utils";
 import { useAuth } from "@/providers/auth-provider";
-import { BadgeCheck, Bell, ChevronsUpDown, LogOut } from "lucide-react";
+import { BadgeCheck, ChevronsUpDown, Import, LogOut } from "lucide-react";
 import { useState } from "react";
+;
 
 export function NavUser() {
   const { isMobile } = useSidebar();
   const { handleLogout } = useAuth();
   const [accountOpen, setAccountOpen] = useState(false);
+  const [importOpen, setImportOpen] = useState(false);
 
   const { data, error, isLoading } = useGetUserQuery();
 
@@ -87,9 +90,10 @@ export function NavUser() {
                   <BadgeCheck />
                   Account
                 </DropdownMenuItem>
-                <DropdownMenuItem>
-                  <Bell />
-                  Notifications
+                <DropdownMenuItem onClick={() => setImportOpen(true)}>
+                  <Import />
+                  {/*https://ui.shadcn.com/docs/components/input*/}
+                  Import Transactions
                 </DropdownMenuItem>
               </DropdownMenuGroup>
               <DropdownMenuSeparator />
@@ -110,6 +114,10 @@ export function NavUser() {
         name={data?.name ?? ""}
         username={data?.username ?? ""}
         profileImage={data?.image ?? ""}
+      />
+      <ImportTransactionPage
+        open={importOpen}
+        setOpen={setImportOpen}
       />
     </>
   );
